+++ client/src/pages/WatchLater/WatchLater.js
@@ -1,14 +1,7 @@
-import { Link } from 'react-router-dom';
-import { Card } from '../../components/Card/Card';
+import { WatchLaterVideos } from '../../components/Sections/WatchLaterVideos';
 import { SidebarDesktop } from '../../components/Sidebar/SidebarDesktop';
-import { useData } from '../../context/DataContext';
-import { handleRemoveWatchLater } from '../../utils/requests';
-import './WatchLater.css';
 
 export const WatchLater = () => {
-   const { state, dispatch } = useData();
-   const { videosWatchLater } = state;
-
    return (
       <>
          <div className='wrapper'>
@@ -16,29 +9,7 @@ export const WatchLater = () => {
                <SidebarDesktop />
             </div>
             <div className='wrapper-videos'>
-               <div className='wrapper-video-list'>
-                  <div className='title'>Watch Later</div>
-                  {videosWatchLater.length > 0 ? (
-                     <div className='video-list'>
-                        {videosWatchLater.map((video) => (
-                           <div key={video.id} className='video-card'>
-                              <div
-                                 onClick={() =>
-                                    handleRemoveWatchLater({ dispatch, video })
-                                 }
-                                 className='video-remove'>
-                                 &times;
-                              </div>
-                              <Link to={`/watch/${video.watchID}`}>
-                                 <Card key={video.id} video={video} />
-                              </Link>
-                           </div>
-                        ))}
-                     </div>
-                  ) : (
-                     <div className='empty'>This list has no videos.</div>
-                  )}
-               </div>
+               <WatchLaterVideos />
             </div>
          </div>
       </>
